remove admin binary
