+++ setup.cfg
@@ -32,4 +32,3 @@ zip_safe = true
 [options.entry_points]
 console_scripts =
     rsconnect = rsconnect.main:cli
-    rsconnect-admin = rsconnect.admin_main:cli
